add rules to modify final warrior attr

+++ 2_sem/metaheu_optim/algo_gen_warriors/include/army.hpp
@@ -8,7 +8,7 @@
 struct ArmyT{
   // Id warrior from warriors pool
   std::set<int> warriors;
-  long score;
+  double score;
 };
 
 struct ArmyComparator{
@@ -18,7 +18,7 @@ struct ArmyComparator{
 };
 
 namespace Army{
-  long Score(const std::set<int> &, const std::vector<Warrior> &);
+  double Score(const std::set<int> &, const std::vector<Warrior> &);
   void Generate(const int, const int, std::set<ArmyT, ArmyComparator> &, const std::vector<Warrior> &); 
   void Crossover(const std::set<int> &, const std::set<int> &, const int, std::set<ArmyT, ArmyComparator> &, const std::vector<Warrior> &);
 }
