+++ 2_sem/metaheu_optim/algo_gen_warriors/include/warrior.hpp
@@ -27,9 +27,9 @@ struct Warrior{
   int control_resis;
   int control;
   int build;
-  int final_attack;
-  int final_defense;
-  int final_shield;
+  double final_attack;
+  double final_defense;
+  double final_shield;
   Warrior(){}
 
   Warrior(int attack, int defense, int shield, int health, int health_reg, int aggressive, int intel, int control_resis, int control, int build) :
